Неужели многопоточка и то, что всё static?
