+++ src/test/java/edu/project3/TestForLogAnalyzer.java
@@ -2,10 +2,8 @@ package edu.project3;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import static edu.project3.TestContentOfReportInLogAnalyzer.thenAssertThatElementInHttpRefererMapExistExpectedTimes;
-import static edu.project3.TestContentOfReportInLogAnalyzer.thenAssertThatElementInHttpUserAgentMapExistExpectedTimes;
-import static edu.project3.TestContentOfReportInLogAnalyzer.thenAssertThatElementInStatusMapExistExpectedTimes;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestForLogAnalyzer {
@@ -46,135 +44,206 @@ public class TestForLogAnalyzer {
     void thenCheckHttpUserAgentMap(int expectedSize, int actualSize) {
         assertEquals(expectedSize, actualSize);
     }
-
-    @Test
-    @DisplayName("Парсинг с сайта")
-    public void testThatGetHttpAndReturnsMetricFile() {
-        //given: http-server & format
-        String[] args = {"--path",
-            "https://raw.githubusercontent.com/elastic/examples/master/Common%20Data%20Formats/nginx_logs/nginx_logs",
-            "--format", "adoc"};
-
-        //when: parse all logs
-        LogAnalyzer.main(args);
-
-        //then: check for same sizes in all maps
-        thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
-        thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
-        thenCheckRequestMap(5, DataClass.getRequestMap().size());
-        thenCheckStatusMap(6, DataClass.getStatusMap().size());
-        thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
-        thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
-        thenCheckRemoteUserMap(1, DataClass.getRemoteUserMap().size());
-        thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
-
-        thenAssertThatElementInHttpUserAgentMapExistExpectedTimes(
-            "Debian APT-HTTP/1.3 (1.0.1ubuntu2)",
-            11830
-        );
-        thenAssertThatElementInHttpUserAgentMapExistExpectedTimes(
-            "Debian APT-HTTP/1.3 (0.9.7.9)",
-            11365
-        );
-
-        //Проверка ReferMap
-        thenAssertThatElementInHttpRefererMapExistExpectedTimes(
-            "http://www.elasticsearch.org/overview/elkdownloads/",
-            6
-        );
-
-        //Проверка statusMap
-        thenAssertThatElementInStatusMapExistExpectedTimes(
-            304,
-            13330
-        );
+    static void thenAssertThatElementInHttpUserAgentMapExistExpectedTimes(String element, int expected){
+        assertEquals(expected,DataClass.getHttpUserAgentMap().get(element));
     }
 
-    @Test
-    @DisplayName("Парсинг с чётко данного пути в системе")
-    public void testThatGetPathToOneFileAndReturnsMetric() {
-        //given: full path
-        String[] args = {"--path", "src/test/java/edu/project3/resources/log1.txt", "--format", "adoc"};
-
-        //when: parse all logs
-        LogAnalyzer.main(args);
-
-        thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
-        thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
-        thenCheckRequestMap(5, DataClass.getRequestMap().size());
-        thenCheckStatusMap(6, DataClass.getStatusMap().size());
-        thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
-        thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
-        thenCheckRemoteUserMap(4, DataClass.getRemoteUserMap().size());
-        thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
-
+    static void thenAssertThatElementInHttpRefererMapExistExpectedTimes(String element, int expected){
+        assertEquals(expected,DataClass.getHttpRefererMap().get(element));
     }
 
-    @Test
-    @DisplayName("Парсинг всего и вся из папки, но указан интервал времени так, что логов нет")
-    public void testThatGetFolderAndReturnedEmptyMaps() {
-        //given: path, format & date interval, but date interval is bad
-        String[] args =
-            {"--path", "src/test/java/edu/project3/resources/", "--format", "markdown", "--from", "2023-08-31", "--to",
-                "2023-09-24"};
-
-        //when: parse all logs
-        LogAnalyzer.main(args);
-
-        //then: check for same sizes in all maps
-        thenCheckTimeLocalMap(0, DataClass.getTimeLocalMap().size());
-        thenCheckRemoteAddrMap(0, DataClass.getRemoteAddrMap().size());
-        thenCheckRequestMap(0, DataClass.getRequestMap().size());
-        thenCheckStatusMap(0, DataClass.getStatusMap().size());
-        thenCheckBodyBytesSentMap(0, DataClass.getBodyBytesSentMap().size());
-        thenCheckHttpRefererMap(0, DataClass.getHttpRefererMap().size());
-        thenCheckRemoteUserMap(0, DataClass.getRemoteUserMap().size());
-        thenCheckHttpUserAgentMap(0, DataClass.getHttpUserAgentMap().size());
-
+    static void thenAssertThatElementInStatusMapExistExpectedTimes(int element, int expected){
+        assertEquals(expected,DataClass.getStatusMap().get(element));
     }
 
-    @Test
-    @DisplayName("Парсинг всего и вся из папки, но интервал позволяет брать логи")
-    public void testThatGetFolderAndReturnedParsedFileAsLogs() {
-        //given: path, format & date interval, but date interval is bad
-        String[] args =
-            {"--path", "src/test/java/edu/project3/resources/", "--format", "markdown", "--from", "2003-08-31", "--to",
-                "2023-09-24"};
+    @Nested
+    class testOfMapSize {
+        @Test
+        @DisplayName("Парсинг с сайта")
+        public void testThatGetHttpAndReturnsMetricFile() {
+            //given: http-server & format
+            String[] args = {"--path",
+                "https://raw.githubusercontent.com/elastic/examples/master/Common%20Data%20Formats/nginx_logs/nginx_logs",
+                "--format", "adoc"};
 
-        //when: parse all logs
-        LogAnalyzer.main(args);
+            //when: parse all logs
+            LogAnalyzer.main(args);
 
-        //then: check for same sizes in all maps
-        thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
-        thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
-        thenCheckRequestMap(5, DataClass.getRequestMap().size());
-        thenCheckStatusMap(6, DataClass.getStatusMap().size());
-        thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
-        thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
-        thenCheckRemoteUserMap(4, DataClass.getRemoteUserMap().size());
-        thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
+            //then: check for same sizes in all maps
+            thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
+            thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
+            thenCheckRequestMap(5, DataClass.getRequestMap().size());
+            thenCheckStatusMap(6, DataClass.getStatusMap().size());
+            thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
+            thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
+            thenCheckRemoteUserMap(1, DataClass.getRemoteUserMap().size());
+            thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
 
+        }
+
+        @Test
+        @DisplayName("Парсинг с чётко данного пути в системе")
+        public void testThatGetPathToOneFileAndReturnsMetric() {
+            //given: full path
+            String[] args = {"--path", "src/test/java/edu/project3/resources/log1.txt", "--format", "adoc"};
+
+            //when: parse all logs
+            LogAnalyzer.main(args);
+
+            thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
+            thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
+            thenCheckRequestMap(5, DataClass.getRequestMap().size());
+            thenCheckStatusMap(6, DataClass.getStatusMap().size());
+            thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
+            thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
+            thenCheckRemoteUserMap(4, DataClass.getRemoteUserMap().size());
+            thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
+
+        }
+
+        @Test
+        @DisplayName("Парсинг всего и вся из папки, но указан интервал времени так, что логов нет")
+        public void testThatGetFolderAndReturnedEmptyMaps() {
+            //given: path, format & date interval, but date interval is bad
+            String[] args =
+                {"--path", "src/test/java/edu/project3/resources/", "--format", "markdown", "--from", "2023-08-31",
+                    "--to",
+                    "2023-09-24"};
+
+            //when: parse all logs
+            LogAnalyzer.main(args);
+
+            //then: check for same sizes in all maps
+            thenCheckTimeLocalMap(0, DataClass.getTimeLocalMap().size());
+            thenCheckRemoteAddrMap(0, DataClass.getRemoteAddrMap().size());
+            thenCheckRequestMap(0, DataClass.getRequestMap().size());
+            thenCheckStatusMap(0, DataClass.getStatusMap().size());
+            thenCheckBodyBytesSentMap(0, DataClass.getBodyBytesSentMap().size());
+            thenCheckHttpRefererMap(0, DataClass.getHttpRefererMap().size());
+            thenCheckRemoteUserMap(0, DataClass.getRemoteUserMap().size());
+            thenCheckHttpUserAgentMap(0, DataClass.getHttpUserAgentMap().size());
+
+        }
+
+        @Test
+        @DisplayName("Парсинг всего и вся из папки, но интервал позволяет брать логи")
+        public void testThatGetFolderAndReturnedParsedFileAsLogs() {
+            //given: path, format & date interval, but date interval is bad
+            String[] args =
+                {"--path", "src/test/java/edu/project3/resources/", "--format", "markdown", "--from", "2003-08-31",
+                    "--to",
+                    "2023-09-24"};
+
+            //when: parse all logs
+            LogAnalyzer.main(args);
+
+            //then: check for same sizes in all maps
+            thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
+            thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
+            thenCheckRequestMap(5, DataClass.getRequestMap().size());
+            thenCheckStatusMap(6, DataClass.getStatusMap().size());
+            thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
+            thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
+            thenCheckRemoteUserMap(4, DataClass.getRemoteUserMap().size());
+            thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
+
+        }
+
+        @Test
+        @DisplayName(
+            "Дан путь через glob. Находим два файла, удовлетворяющие условию, считываем. Интервал времени хороший")
+        public void testThatGetFolderWithMissingsAndFinalFileNameAndReturnedParsedFileAsLogs() {
+            //given: path, format & date interval, but date interval is good
+            String[] args =
+                {"--path", "src/test/java/edu/project3/resources/**/log1.txt", "--from", "2000-08-31", "--format",
+                    "markdown", "--to", "2029-09-24"};
+
+            //when: parse all logs
+            LogAnalyzer.main(args);
+
+            //then: check for same sizes in all maps
+            thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
+            thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
+            thenCheckRequestMap(5, DataClass.getRequestMap().size());
+            thenCheckStatusMap(6, DataClass.getStatusMap().size());
+            thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
+            thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
+            thenCheckRemoteUserMap(4, DataClass.getRemoteUserMap().size());
+            thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
+        }
     }
 
-    @Test
-    @DisplayName("Дан путь через glob. Находим два файла, удовлетворяющие условию, считываем. Интервал времени хороший")
-    public void testThatGetFolderWithMissingsAndFinalFileNameAndReturnedParsedFileAsLogs() {
-        //given: path, format & date interval, but date interval is good
-        String[] args =
-            {"--path", "src/test/java/edu/project3/resources/**/log1.txt", "--from", "2000-08-31", "--format",
-                "markdown", "--to", "2029-09-24"};
+    @Nested
+    class testOfContent {
 
-        //when: parse all logs
-        LogAnalyzer.main(args);
+        @Test
+        @DisplayName("Парсинг с сайта")
+        public void testThatGetHttpAndReturnsContentOfMetricFile() {
+            //given: http-server & format
+            String[] args = {"--path",
+                "https://raw.githubusercontent.com/elastic/examples/master/Common%20Data%20Formats/nginx_logs/nginx_logs",
+                "--format", "adoc"};
 
-        //then: check for same sizes in all maps
-        thenCheckTimeLocalMap(22467, DataClass.getTimeLocalMap().size());
-        thenCheckRemoteAddrMap(2660, DataClass.getRemoteAddrMap().size());
-        thenCheckRequestMap(5, DataClass.getRequestMap().size());
-        thenCheckStatusMap(6, DataClass.getStatusMap().size());
-        thenCheckBodyBytesSentMap(187, DataClass.getBodyBytesSentMap().size());
-        thenCheckHttpRefererMap(8, DataClass.getHttpRefererMap().size());
-        thenCheckRemoteUserMap(4, DataClass.getRemoteUserMap().size());
-        thenCheckHttpUserAgentMap(136, DataClass.getHttpUserAgentMap().size());
+            //when: parse all logs
+            LogAnalyzer.main(args);
+
+            //then: check for some basic content in maps
+
+            // Проверка основного содержимого карты HTTP_USER_AGENT_MAP
+            thenAssertThatElementInHttpUserAgentMapExistExpectedTimes(
+                "Debian APT-HTTP/1.3 (1.0.1ubuntu2)",
+                11830
+            );
+            thenAssertThatElementInHttpUserAgentMapExistExpectedTimes(
+                "Debian APT-HTTP/1.3 (0.9.7.9)",
+                11365
+            );
+
+            //Проверка ReferMap
+            thenAssertThatElementInHttpRefererMapExistExpectedTimes(
+                "http://www.elasticsearch.org/overview/elkdownloads/",
+                6
+            );
+
+            //Проверка statusMap
+            thenAssertThatElementInStatusMapExistExpectedTimes(
+                304,
+                13330
+            );
+        }
+
+        @Test
+        @DisplayName("Парсинг с чётко данного пути в системе")
+        public void testThatGetPathToOneFileAndReturnsContentOfMetric() {
+            //given: full path
+            String[] args = {"--path", "src/test/java/edu/project3/resources/log1.txt", "--format", "adoc"};
+
+            //when: parse all logs
+            LogAnalyzer.main(args);
+
+            //then: check for some basic content in maps
+
+            // Проверка основного содержимого карты HTTP_USER_AGENT_MAP
+            thenAssertThatElementInHttpUserAgentMapExistExpectedTimes(
+                "Debian APT-HTTP/1.3 (1.0.1ubuntu2)",
+                11830
+            );
+            thenAssertThatElementInHttpUserAgentMapExistExpectedTimes(
+                "Debian APT-HTTP/1.3 (0.9.7.9)",
+                11365
+            );
+
+            //Проверка ReferMap
+            thenAssertThatElementInHttpRefererMapExistExpectedTimes(
+                "http://www.elasticsearch.org/overview/elkdownloads/",
+                6
+            );
+
+            //Проверка statusMap
+            thenAssertThatElementInStatusMapExistExpectedTimes(
+                304,
+                13334
+            );
+        }
     }
 }
